pyproject.toml
[project]
name = "altspell"
version = "0.1.0"
description = "A flask web API for translating between traditional English spellings and alternative spellings"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name="Nicholas Johnson", email="njohnson@posteo.net" },
]
classifiers = [
  "Private :: Do Not Upload",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "License-Expression :: GPL-3.0-or-later",
  "Operating System :: OS Independent",
]
dependencies = [
    "dependency_injector",
    "flask",
    "flask_sqlalchemy",
    "flask_cors",
    "psycopg2",
    "pytz"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "altspell"
